WIP: first very rough draft of server core builder. Doesn't meet parity with old functional approach yet (slower / unstable?).
