+++ mistralrs-server-core/src/handlers.rs
@@ -1,0 +1,62 @@
+use anyhow::Result;
+use axum::extract::{Json, State};
+use mistralrs_core::{parse_isq_value, MistralRs, Request};
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
+
+use crate::{
+    openai::{ModelObject, ModelObjects},
+    ExtractedMistralState,
+};
+
+#[utoipa::path(
+  get,
+  tag = "Mistral.rs",
+  path = "/v1/models",
+  responses((status = 200, description = "Served model info", body = ModelObjects))
+)]
+pub async fn models(State(state): ExtractedMistralState) -> Json<ModelObjects> {
+    Json(ModelObjects {
+        object: "list",
+        data: vec![ModelObject {
+            id: state.get_id(),
+            object: "model",
+            created: state.get_creation_time(),
+            owned_by: "local",
+        }],
+    })
+}
+
+#[utoipa::path(
+  get,
+  tag = "Mistral.rs",
+  path = "/health",
+  responses((status = 200, description = "Server is healthy"))
+)]
+pub async fn health() -> &'static str {
+    "OK"
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
+pub struct ReIsqRequest {
+    #[schema(example = "Q4K")]
+    ggml_type: String,
+}
+
+#[utoipa::path(
+  post,
+  tag = "Mistral.rs",
+  path = "/re_isq",
+  request_body = ReIsqRequest,
+  responses((status = 200, description = "Reapply ISQ to a non GGUF or GGML model."))
+)]
+pub async fn re_isq(
+    State(state): ExtractedMistralState,
+    Json(request): Json<ReIsqRequest>,
+) -> Result<String, String> {
+    let repr = format!("Re ISQ: {:?}", request.ggml_type);
+    MistralRs::maybe_log_request(state.clone(), repr.clone());
+    let request = Request::ReIsq(parse_isq_value(&request.ggml_type, None)?);
+    state.get_sender().unwrap().send(request).await.unwrap();
+    Ok(repr)
+}
